# docs/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Outlier API",
  "description": "유튜브 다채널 유해댓글 관리 서비스",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/channels/{channel_id}/queue": {
   "get": {
    "tags": [
     "review"
    ],
    "summary": "검토 큐 (관리자가 판단해야 할 댓글)",
    "operationId": "queue_api_channels__channel_id__queue_get",
    "parameters": [
     {
      "name": "channel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Channel Id"
      }
     },
     {
      "name": "kind",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "all",
        "judge",
        "info"
       ],
       "type": "string",
       "description": "judge=판단 필요만 / info=참고만",
       "default": "all",
       "title": "Kind"
      },
      "description": "judge=판단 필요만 / info=참고만"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/QueueItem"
         },
         "title": "Response Queue Api Channels  Channel Id  Queue Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/channels/{channel_id}/hidden": {
   "get": {
    "tags": [
     "review"
    ],
    "summary": "숨김 목록 (오탐을 발견하는 통로)",
    "operationId": "hidden_api_channels__channel_id__hidden_get",
    "parameters": [
     {
      "name": "channel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Channel Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/QueueItem"
         },
         "title": "Response Hidden Api Channels  Channel Id  Hidden Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/comments/{comment_id}/action": {
   "post": {
    "tags": [
     "review"
    ],
    "summary": "조치 (숨김 / 유지 / 채널차단)",
    "operationId": "act_api_comments__comment_id__action_post",
    "parameters": [
     {
      "name": "comment_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Comment Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ActionRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ActionResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/channels/{channel_id}/stats": {
   "get": {
    "tags": [
     "review"
    ],
    "summary": "채널 처리 현황",
    "operationId": "stats_api_channels__channel_id__stats_get",
    "parameters": [
     {
      "name": "channel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Channel Id"
      }
     },
     {
      "name": "period",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "today",
        "7d",
        "30d",
        "all"
       ],
       "type": "string",
       "description": "수집 시각 기준",
       "default": "all",
       "title": "Period"
      },
      "description": "수집 시각 기준"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Stats"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/channels/{channel_id}/history": {
   "get": {
    "tags": [
     "review"
    ],
    "summary": "처리 이력 (누가 언제 무엇을 왜)",
    "description": "조치 이력을 최신순으로. 관리자가 자기 판단을 되짚어볼 수 있어야 하고,\n나중에 이 이력이 채널별 맥락을 만드는 재료가 된다 (F_R_116).",
    "operationId": "history_api_channels__channel_id__history_get",
    "parameters": [
     {
      "name": "channel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Channel Id"
      }
     },
     {
      "name": "action",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "all",
        "hide",
        "keep",
        "ban_author"
       ],
       "type": "string",
       "default": "all",
       "title": "Action"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/HistoryItem"
         },
         "title": "Response History Api Channels  Channel Id  History Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/comments/{comment_id}/similar": {
   "get": {
    "tags": [
     "review"
    ],
    "summary": "과거 유사 사례 (F_R_114)",
    "description": "뜻이 비슷한 과거 댓글과, 그때 관리자가 내린 결정을 함께 돌려준다.\n\n단어가 겹치는 댓글이 아니라 '같은 사안'을 찾는 게 목적이다 —\n\"주소 알아내서 찾아간다\"와 \"어디 사는지 알아냈음\"은 겹치는 단어가\n없지만 관리자에겐 같은 건이다.\n\n같은 채널 안에서만 찾는다. 채널(고객사)간 데이터 결합·비교 노출은\nYouTube API 정책상 금지다.",
    "operationId": "similar_api_comments__comment_id__similar_get",
    "parameters": [
     {
      "name": "comment_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Comment Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 10,
       "minimum": 1,
       "default": 3,
       "title": "Limit"
      }
     },
     {
      "name": "only_reviewed",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "관리자가 실제로 처리한 것만 보기",
       "default": false,
       "title": "Only Reviewed"
      },
      "description": "관리자가 실제로 처리한 것만 보기"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/SimilarCase"
         },
         "title": "Response Similar Api Comments  Comment Id  Similar Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
   "ActionRequest": {
    "properties": {
     "action": {
      "type": "string",
      "enum": [
       "hide",
       "keep",
       "ban_author"
      ],
      "title": "Action",
      "description": "hide=숨김 / keep=유지 / ban_author=채널차단"
     },
     "actor": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100
       },
       {
        "type": "null"
       }
      ],
      "title": "Actor",
      "description": "처리한 관리자"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note",
      "description": "판단 근거 메모"
     }
    },
    "type": "object",
    "required": [
     "action"
    ],
    "title": "ActionRequest"
   },
   "ActionResult": {
    "properties": {
     "comment_id": {
      "type": "integer",
      "title": "Comment Id"
     },
     "action": {
      "type": "string",
      "enum": [
       "hide",
       "keep",
       "ban_author"
      ],
      "title": "Action"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "youtube_synced": {
      "type": "boolean",
      "title": "Youtube Synced",
      "description": "유튜브 실제 반영 여부. 채널 소유자 OAuth 가 없어 항상 False",
      "default": false
     },
     "reviewed_at": {
      "type": "string",
      "format": "date-time",
      "title": "Reviewed At"
     }
    },
    "type": "object",
    "required": [
     "comment_id",
     "action",
     "status",
     "reviewed_at"
    ],
    "title": "ActionResult"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "HistoryItem": {
    "properties": {
     "action_id": {
      "type": "integer",
      "title": "Action Id"
     },
     "comment_id": {
      "type": "integer",
      "title": "Comment Id"
     },
     "action": {
      "type": "string",
      "title": "Action",
      "description": "hide=숨김 / keep=유지 / ban_author=채널차단"
     },
     "actor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Actor"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "youtube_synced": {
      "type": "boolean",
      "title": "Youtube Synced"
     },
     "executed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Executed At"
     },
     "content": {
      "type": "string",
      "title": "Content"
     },
     "author_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Author Name"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category",
      "description": "조치 당시 AI 가 본 분류"
     },
     "reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason"
     }
    },
    "type": "object",
    "required": [
     "action_id",
     "comment_id",
     "action",
     "actor",
     "note",
     "youtube_synced",
     "executed_at",
     "content",
     "author_name",
     "reason"
    ],
    "title": "HistoryItem"
   },
   "QueueItem": {
    "properties": {
     "comment_id": {
      "type": "integer",
      "title": "Comment Id"
     },
     "severity": {
      "type": "string",
      "enum": [
       "critical",
       "high",
       "medium",
       "low"
      ],
      "title": "Severity",
      "description": "처리 우선순위. 카테고리 + 확산도로 정한다"
     },
     "channel_title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Channel Title"
     },
     "video_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Video Id"
     },
     "video_title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Video Title"
     },
     "published_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Published At",
      "description": "유튜브에 달린 시각"
     },
     "content": {
      "type": "string",
      "title": "Content"
     },
     "author_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Author Name"
     },
     "like_count": {
      "type": "integer",
      "title": "Like Count"
     },
     "total_reply_count": {
      "type": "integer",
      "title": "Total Reply Count"
     },
     "is_reply": {
      "type": "boolean",
      "title": "Is Reply"
     },
     "parent_content": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent Content",
      "description": "답글이면 부모 댓글 본문"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "destination": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Destination",
      "description": "queue_judge=판단 필요 / queue_info=참고"
     },
     "decided_by": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Decided By",
      "description": "rule=관리자 등록어 / llm"
     },
     "label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason"
     },
     "rule_value": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rule Value",
      "description": "등록어에 걸렸으면 그 단어"
     }
    },
    "type": "object",
    "required": [
     "comment_id",
     "severity",
     "channel_title",
     "video_id",
     "video_title",
     "content",
     "author_name",
     "like_count",
     "total_reply_count",
     "is_reply",
     "status",
     "label",
     "category",
     "reason"
    ],
    "title": "QueueItem"
   },
...
   "SimilarCase": {
    "properties": {
     "comment_id": {
      "type": "integer",
      "title": "Comment Id"
     },
     "content": {
      "type": "string",
      "title": "Content"
     },
     "similarity": {
      "type": "number",
      "title": "Similarity",
      "description": "1.0 에 가까울수록 비슷하다"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "status": {
      "type": "string",
      "title": "Status",
      "description": "passed / queued / hidden"
     },
     "reviewed": {
      "type": "boolean",
      "title": "Reviewed",
      "description": "관리자가 실제로 처리한 건인지"
     },
     "action": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Action",
      "description": "hide / keep / ban_author"
     },
     "actor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Actor"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "executed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Executed At"
     }
    },
    "type": "object",
    "required": [
     "comment_id",
     "content",
     "similarity",
     "category",
     "status",
     "reviewed",
     "actor",
     "note",
     "executed_at"
    ],
    "title": "SimilarCase"
   },
   "Stats": {
    "properties": {
     "channel_id": {
      "type": "integer",
      "title": "Channel Id"
     },
     "period": {
      "type": "string",
      "title": "Period"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "pending": {
      "type": "integer",
      "title": "Pending"
     },
     "passed": {
      "type": "integer",
      "title": "Passed"
     },
     "queued": {
      "type": "integer",
      "title": "Queued"
     },
     "hidden": {
      "type": "integer",
      "title": "Hidden"
     },
     "unreviewed": {
      "type": "integer",
      "title": "Unreviewed",
      "description": "큐에 남아 관리자를 기다리는 건수"
     },
     "review_rate": {
      "type": "number",
      "title": "Review Rate",
      "description": "검토 전환율. NF_R_104 목표 30% 이하"
     },
     "by_category": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "By Category"
     }
    },
    "type": "object",
    "required": [
     "channel_id",
     "period",
     "total",
     "pending",
     "passed",
     "queued",
     "hidden",
     "unreviewed",
     "review_rate",
     "by_category"
    ],
    "title": "Stats"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}